Makes all logic models implement Serializable
Makes all models implement Serializable so they can be transferred with
object streams. This commit also adds a simple test of placing a ship on
the board and sending it to the server.

Fixes a bug where some references to x and y were the wrong way round.
Also adds a getter in Board to get ArrayList of ships.