Modifications to app logic and messages
Modifies Square to have 3 states, where it either contains a ship or
does not, or is unknown. The unknown case if for the opponents board.
This commit also adds an update method to Square, to update the state of
Square based on if the guess hit a ship and if it sunk a ship. As a ship
is returned if a ship is sunk, I added a sink method to Ship which sets
its health to 0, which the client would do in response to the server
telling it has sunk.

This commit also modifies MoveResponseMessage, so the Ship that it
contains is the ship that has sunk, not the ship that is hit. This is to
not reveal information about a ship that has not been sunk to the
opponent. It now contains a boolean, stating whether the move hit ship
or not. The coordinates are now stored as two separate integers.

A method applyMove is added to Board, which takes the above mentioned
MoveResponseMessage as parameter, and updates the state of the board.