Adds ability to send notifications to client
Adds a new class NotificationMessage, with public final static ints
representing different messages and ability to hold Strings. This allows
notifying clients of game events, such as who's turn it is and the
result of a game, and sending information like the player's opponents
name.

This commit also added some incomplete logic to the server Game class to
handle a game between two players. Other logic models were slightly
adjusted to achieve this.

MoveMessage is another new class added, which contains coordinates of
the square a player wishes to guess. A boolean was added to
MoveResponseMessage to state if the response is targeted for the
player's board or their opponent's.

Instances of the Client test can now battle each other by guessing
random squares.