Sends more notifications to clients
Game.java now sends more notifications to handle a few unlikely edge
cases. These include notifying a client if the player tries to make a
move when it is not their turn, submitting a move guessing a square that
is not on the grid and guessing a square they have already guessed. All
of these should be rejected by the GUI too.

Clients are notified if either player took too long to make a move, and
are told the resulting outcome of the game. Taking too long to make a
move causes the game to end.

A class called TestBoards is added, which contains multiple boards the
test client can choose from, so two test clients can play against
each other with a different board.