Makes model responsible for ship creation
Removes the code from BoardView that creates its own ships, and instead
gets them from the ships already made by the Board model. This behaviour
was causing the server board validation to fail, as the board contained
extra ships with no squares assigned.

Adds INVALID_BOARD notification message, which is sent to a client that
submits an invalid board. NOT_IN_GAME is also added, which is sent to a
client trying to submit a board when they are not playing a game.