Uses ChangeEvent and draws ship when sank
Changes from using PropertyChangeEvent and PropertyChangeListener to
using ChangeEvent and ChangeListener for the Square view and model.

Updates references of squares in ship object received from server to
point to squares on the board that contains it. Also paints the ship on
the opponents board when a player sinks the ship.