Fixes handling moves from out of turn players
Fixes a bug where NOT_YOUR_TURN was sent to the player who's turn it was
instead of the player who sent the move out of turn. It was also
possible the notification was attempted to be sent to a null object,
causing an exception. This occurred when a player sent a move when still
in ship placement phase.

Adds more NotificationMessages to the chat in the game client for easier
debugging.