Stores opponent name Client, prevents playing self
Also refactors code where NotificationMessages are sent to prevent
repetition.

When an opponent's name is received, it is stored in a private String in
Client.

The server now checks to make sure a player is not sending a request to
join themselves, and will send an error if they attempt to do so.

Creates a method in Player, writeNotification(), which initializes a new
NotificationMessage object and sends it to the client.